# AES for ARM32 without using crypto hardware instructions