… wolfcrypt/src/port/arm/armv8-32-aes-asm.S
… wolfcrypt/src/port/arm/armv8-32-curve25519.S
… wolfcrypt/src/port/arm/armv8-32-curve25519_c.c
… wolfcrypt/src/port/arm/armv8-32-sha256-asm.S
… wolfcrypt/src/port/arm/armv8-32-sha256-asm_c.c
… wolfcrypt/src/port/arm/armv8-32-sha512-asm.S
… wolfcrypt/src/port/arm/armv8-32-sha512-asm_c.c
… wolfcrypt/src/port/arm/armv8-aes.c