# Description

AES-ECB, AES-CBC, AES-CTR, AES-GCM, AES-CCM
(ARM32, No NEON)

# Testing

Tested with QEMU and on Linaro board.

# Checklist

 - [ ] added tests
 - [ ] updated/added doxygen
 - [ ] updated appropriate READMEs
 - [ ] Updated manual and documentation