@haydenroche5 2022-10-13
I get this error when building with `--enable-armasm` on my Raspberry Pi:

```
make -j5  all-recursive
make[1]: warning: -j5 forced in submake: resetting jobserver mode.
make[1]: Entering directory '/home/pi/wolfssl'
make[2]: Entering directory '/home/pi/wolfssl'
make[2]: warning: -j5 forced in submake: resetting jobserver mode.
  CPPAS    wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-32-sha256-asm.lo
  CPPAS    wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-32-sha512-asm.lo
  CC       wolfcrypt/benchmark/benchmark.o
  CC       wolfcrypt/src/src_libwolfssl_la-hmac.lo
  CC       wolfcrypt/src/src_libwolfssl_la-hash.lo
  CC       wolfcrypt/src/src_libwolfssl_la-cpuid.lo
  CC       wolfcrypt/src/src_libwolfssl_la-kdf.lo
  CC       wolfcrypt/src/src_libwolfssl_la-random.lo
  CC       wolfcrypt/src/src_libwolfssl_la-sha256.lo
  CC       wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-sha256.lo
  CC       wolfcrypt/src/src_libwolfssl_la-rsa.lo
  CC       wolfcrypt/src/src_libwolfssl_la-sp_int.lo
  CC       wolfcrypt/src/src_libwolfssl_la-aes.lo
  CC       wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-aes.lo
  CC       wolfcrypt/src/src_libwolfssl_la-sha.lo
  CC       wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-sha512.lo
  CPPAS    wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-sha512-asm.lo
  CC       wolfcrypt/src/src_libwolfssl_la-logging.lo
  CC       wolfcrypt/src/src_libwolfssl_la-wc_port.lo
  CC       wolfcrypt/src/src_libwolfssl_la-error.lo
  CC       wolfcrypt/src/src_libwolfssl_la-wc_encrypt.lo
  CC       wolfcrypt/src/src_libwolfssl_la-signature.lo
  CC       wolfcrypt/src/src_libwolfssl_la-wolfmath.lo
  CC       wolfcrypt/src/src_libwolfssl_la-memory.lo
  CC       wolfcrypt/src/src_libwolfssl_la-dh.lo
  CC       wolfcrypt/src/src_libwolfssl_la-asn.lo
  CC       wolfcrypt/src/src_libwolfssl_la-coding.lo
  CC       wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-poly1305.lo
  CC       wolfcrypt/src/src_libwolfssl_la-poly1305.lo
  CC       wolfcrypt/src/src_libwolfssl_la-md5.lo
  CC       wolfcrypt/src/src_libwolfssl_la-pwdbased.lo
  CC       wolfcrypt/src/src_libwolfssl_la-pkcs12.lo
  CC       wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-chacha.lo
  CC       wolfcrypt/src/src_libwolfssl_la-chacha20_poly1305.lo
  CC       wolfcrypt/src/src_libwolfssl_la-ecc.lo
/tmp/cc1tgIXA.s: Assembler messages:
/tmp/cc1tgIXA.s:241: Error: first transfer register must be even -- `ldrd r11,r10,[r14,#4*14]'
make[2]: *** [Makefile:6216: wolfcrypt/src/port/arm/src_libwolfssl_la-armv8-chacha.lo] Error 1
make[2]: *** Waiting for unfinished jobs....
make[2]: Leaving directory '/home/pi/wolfssl'
make[1]: *** [Makefile:7150: all-recursive] Error 1
make[1]: Leaving directory '/home/pi/wolfssl'
make: *** [Makefile:4222: all] Error 2
```

CPU specs:

```
Architecture:        armv7l
Byte Order:          Little Endian
CPU(s):              4
On-line CPU(s) list: 0-3
Thread(s) per core:  1
Core(s) per socket:  4
Socket(s):           1
Vendor ID:           ARM
Model:               3
Model name:          Cortex-A72
Stepping:            r0p3
CPU max MHz:         1500.0000
CPU min MHz:         600.0000
BogoMIPS:            108.00
Flags:               half thumb fastmult vfp edsp neon vfpv3 tls vfpv4 idiva idivt vfpd32 lpae evtstrm crc32
```