@SparkiDev 2022-10-14
Fixed ldrd and strd usage.