@haydenroche5 2022-10-21
Seeing a nice speedup!

Before:

```
------------------------------------------------------------------------------
 wolfSSL version 5.5.1
------------------------------------------------------------------------------
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
AES-128-GCM-enc             30 MiB took 1.042 seconds,   28.788 MiB/s
AES-128-GCM-dec             35 MiB took 1.159 seconds,   30.208 MiB/s
AES-192-GCM-enc             30 MiB took 1.100 seconds,   27.261 MiB/s
AES-192-GCM-dec             30 MiB took 1.099 seconds,   27.294 MiB/s
AES-256-GCM-enc             25 MiB took 1.011 seconds,   24.737 MiB/s
AES-256-GCM-dec             25 MiB took 1.010 seconds,   24.757 MiB/s
GMAC Table 4-bit            75 MiB took 1.006 seconds,   74.523 MiB/s
Benchmark complete
```

After:

```
------------------------------------------------------------------------------
 wolfSSL version 5.5.1
------------------------------------------------------------------------------
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
AES-128-GCM-enc             45 MiB took 1.101 seconds,   40.886 MiB/s
AES-128-GCM-dec             45 MiB took 1.060 seconds,   42.470 MiB/s
AES-192-GCM-enc             40 MiB took 1.027 seconds,   38.958 MiB/s
AES-192-GCM-dec             40 MiB took 1.026 seconds,   38.991 MiB/s
AES-256-GCM-enc             40 MiB took 1.111 seconds,   36.018 MiB/s
AES-256-GCM-dec             40 MiB took 1.110 seconds,   36.023 MiB/s
GMAC Table 4-bit            82 MiB took 1.004 seconds,   81.690 MiB/s
Benchmark complete
```